// src/types/Theme.ts
export type ThemeColors = {
    background: string;
    text: string;
    border: string;
    icon: string;
    overlay: string;
    primary: string;
    secondary: string;
    tertiary: string;
    accent: string;
    highlight: string;
  };
  
export type ThemeType = {
    light: ThemeColors;
    dark: ThemeColors;
  };